+++ savu/plugins/reconstructions/astra_recons/base_astra_recon.py
@@ -23,6 +23,8 @@ import astra
 import numpy as np
 
 from savu.plugins.reconstructions.base_recon import BaseRecon
+from savu.core.iterate_plugin_group_utils import enable_iterative_loop, \
+    check_if_end_plugin_in_iterate_group
 
 class BaseAstraRecon(BaseRecon):
 
@@ -30,6 +32,21 @@ class BaseAstraRecon(BaseRecon):
         super(BaseAstraRecon, self).__init__(name)
         self.res = False
 
+    # total number of output datasets
+    def nOutput_datasets(self):
+        if check_if_end_plugin_in_iterate_group(self.exp):
+            return 2
+        else:
+            return 1
+
+    # total number of output datasets that are clones
+    def nClone_datasets(self):
+        if check_if_end_plugin_in_iterate_group(self.exp):
+            return 1
+        else:
+            return 0
+
+    @enable_iterative_loop
     def setup(self):
         self.alg = self.parameters['algorithm']
         self.get_max_frames = self._get_multiple if '3D' in self.alg else self._get_single
@@ -38,7 +55,12 @@ class BaseAstraRecon(BaseRecon):
         out_dataset = self.get_out_datasets()
 
         # if res_norm is required then setup another output dataset
-        if len(out_dataset) == 2:
+        if len(out_dataset) == 3 and self.nClone_datasets() == 1:
+            err_str = "The res_norm output dataset has not yet been " \
+                "implemented for when AstraReconCpu is at the end of an " \
+                "iterative loop"
+            raise ValueError(err_str)
+        elif len(out_dataset) == 2 and self.nClone_datasets() == 0:
             self.res = True
             out_pData = self.get_plugin_out_datasets()
             in_data = self.get_in_datasets()[0]