+++ scripts/plugin_generator/template_elements/parameter_definition.py
@@ -3,13 +3,13 @@
         parameter_name_1:
             visibility: basic
             dtype: int
-            description: "Describe your parameter"
+            description: Describe your parameter
             default: 1
 
         parameter_name_2:
             visibility: intermediate
             dtype: str
-            description: "Describe your parameter"
+            description: Describe your parameter
             default: A default value
             options: [A default value, option1, option2, option3]
             dependency:
@@ -17,7 +17,6 @@
         """
 
     def citation1(self):
-        # use a u notation for unicode characters - for example, mew
         """
         A description of the citation
         bibtex: