+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 All notable changes to this project are documented in this file.
 *******************************************************************
-# Savu Version 4.2, *planned release first quarter of 2022*
+# Savu Version 4.2, *planned release March 2022*
 
 ## _Core_
 * Statistics (TODO)
-* Iterative plugins (TODO)
+* Iterative plugins (A capability to enable some plugins to be iterative):
+  - `iterate` command to enable control over iterative plugins (see `iterate -h` for help)
 
 ## _Existing Plugins_
 
@@ -17,10 +18,11 @@ All notable changes to this project are documented in this file.
 ### Reconstruction
   * AstraReconGPU, 3D GPU methods are added (BP3D_CUDA, CGLS3D_CUDA, SIRT3D_CUDA)
   * FBP3D_CUDA method added (filtering before backprojection with BP3D_CUDA)
-  * ForwardProjector works with 3D geometry
+  * ForwardProjector works with the 3D geometry
   * 3D geometries can accept metadata for x-y shifts and correct the misalignment
   * GPU memory usage check for *tomobar_recon_3D* plugin to avoid CUDA error
   * *tomobar_recon_3D* access to regularisation using Wavelets, try set regularisation method e.g. to 'PD_TV_WAVELETS'
+  * SWLS, PWLS methods for data fidelities are enabled in *tomobar_recon_3D*
 
 ### Filters
   * GPU memory usage check for *ccpi_denoising_gpu_3D* plugin to avoid CUDA error
@@ -32,22 +34,39 @@ All notable changes to this project are documented in this file.
   * *wavelet_denoising_gpu* - a GPU plugin for denoising using Wavelets. Highly optimised for GPU performance.
 ### Corrections
   * *phase_unwrapping* - a plugin for unwrapping phase-retrieved images
+### Centering
+  * *360_centering* - a plugin to calculate centre of rotation. 
 
 ## _Updated and new packages as dependencies_
   * A new [pypwt](https://github.com/pierrepaleo/pypwt "pypwt") GPU wavelet package added through Jenkins build and savu-dep channel
   * ToMoBAR and TomoPhantom packages have been updated
 
-## _Configurator_
+## _Configurator_ 
+  *  Allow a list as a single dimension input to the preview parameter
+  *  Allow a start keyword inside the preview parameter
+  *  Allow a parameter to be a directory path within the Savu folder
+  *  Asterix line added to indicate the start and end of the process list
+  *  Line separator added to indicate an iterative plugin loop
+
+### New Commands
+  * *savu_mod* - a way to modify one parameter present in a plugin list. 
 
 ## _Documentation_
+  *  Plugin API moved to dropdown boxes on every plugin documentation page
+  *  Plugin template links updated
 
 ## _BUGS_
   *  res_norm bug when using AstaReconGPU with CGLS_CUDA
+  *  Fix indentation for the plugin_generator command
+  *  Allow saving to inner plugin directories when using the plugin_generator command
 
 ## Other
   * The test dataset 24737.nxs has been changed to tomo-standard.nxs
-  * The synthetic test data has been added
+  * The synthetic test data has been added 
   * Environment variable *type* is replaced with *GPUarch_nodes*
+  * *savu_mod* - Modify one parameter present in a plugin list. 
+  * Save the job command to a log file
+  * Save the directory the command was run to a log file